requirements.txt
Django>=5.0
gunicorn
psycopg2-binary
python-decouple
python-telegram-bot==20.0
requests